* commit '2de8a864922c64c1e3bac25a799c0026f9a27adb':
  add top level decorator interface for extensible use in downstream consumers
  update parsers to support optional in all typed returns
  additions for use in TDP parsing
  interim

+++ boogie-core/src/main/java/org/mitre/tdp/boogie/Fix.java
@@ -14,5 +14,5 @@ public interface Fix extends HasPosition, Infrastructure {
   /**
    * The elevation in feet of the fix (if a navaid).
    */
-  Optional<Float> elevation();
+  Optional<Double> elevation();
 }
